Refactor authorize form
